requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langserve>=0.3.0
sse-starlette>=3.0.0
pydantic-settings>=2.5.2
streamlit>=1.39.0
chromadb==0.5.4
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
transformers>=4.44.0
torch>=2.3.0
pypdf>=4.3.1
pymupdf>=1.24.10
pdfplumber>=0.11.4
python-docx>=1.1.2
python-multipart>=0.0.9
requests>=2.32.3
numpy>=1.26.4
pytest>=8.3.0
